+++ src/Config/Config.php
@@ -75,6 +75,7 @@ class Config
         'multibanco' => ['pt'],
         'mbway' => ['pt'],
         'swish' => ['se'],
+        'bizum' => ['es', 'ad'],
         'vippsmobilepay' => ['no', 'dk', 'fi'],
     ];
 
@@ -150,6 +151,7 @@ class Config
     const MOLLIE_MAIL_WHEN_EXPIRED = 'MOLLIE_MAIL_WHEN_EXPIRED';
     const MOLLIE_MAIL_WHEN_REFUNDED = 'MOLLIE_MAIL_WHEN_REFUNDED';
     const MOLLIE_MAIL_WHEN_CHARGEBACK = 'MOLLIE_MAIL_WHEN_CHARGEBACK';
+    const MOLLIE_MAIL_WHEN_FAILED = 'MOLLIE_MAIL_WHEN_FAILED';
     // NOTE: const below is needed to check in configuration table if email should be sent. IDE doesn't detect its usage
     const MOLLIE_MAIL_WHEN_PARTIAL_REFUND = 'MOLLIE_MAIL_WHEN_PARTIAL_REFUND';
     const PARTIAL_REFUND_CODE = 'partial_refund';
@@ -342,6 +344,7 @@ class Config
         'paybybank' => 'Pay by Bank',
         'multibanco' => 'Multibanco',
         'mbway' => 'MBway',
+        'bizum' => 'Bizum',
         'vippsmobilepay' => 'Vipps',
     ];